Eight Queens puzzle
The goal of this puzzle is to find a way to arrange 8 queens on a chess
board such that no two queens can run into each other in any one move.
To accomplish this I have allowed users to put queens on the board with
a left click, remove them with a right click, and clear the board.
There is also a hint view which makes it easier for the user to see
valid moves. If you need a solution for testing purposes there are a
few here: http://en.wikipedia.org/wiki/Eight_queens_puzzle.